+++ __tests__/calculate-wprs.test.ts
@@ -1,3 +1,4 @@
+import { describe, it, expect } from "bun:test";
 import { calculateWPRS } from "@/utils/calculate-wprs";
 
 // Use test database url with dummy data only for this test
@@ -5,7 +6,7 @@ process.env.DATABASE_URL = process.env.TEST_DATABASE_URL;
 
 describe("Calculate WPRS", () => {
   const timeout = 200000;
-  it(
+  it.todo(
     "should find the correct forecast for the 120 best pilots in the world",
     async () => {
       const expectedWprs = 120;
@@ -23,7 +24,7 @@ describe("Calculate WPRS", () => {
     timeout,
   );
 
-  it(
+  it.todo(
     "should find the correct forecast for the 60 best pilots in the world (devalued by avg pilots)",
     async () => {
       const expectedWprs = 88.9;
@@ -41,7 +42,7 @@ describe("Calculate WPRS", () => {
     timeout,
   );
 
-  it(
+  it.todo(
     "should find the correct forecast for the 61th to 120th best pilots in the world",
     async () => {
       const expectedWprs = 86.2;