+++ __tests__/get-airtribune-comp.test.ts
@@ -1,3 +1,5 @@
+import { describe, it, expect } from "bun:test";
+
 import { getAirtribuneComp } from "@/utils/get-airtribune-comp";
 
 describe("Airtribune", () => {
@@ -8,8 +10,8 @@ describe("Airtribune", () => {
     const url = "https://airtribune.com/pre-world-cup-reunion-island-2023/info";
     const res = await getAirtribuneComp(url);
 
-    expect(res?.compDate.startDate?.toISOString()).toBe(expectedStartDate);
-    expect(res?.compDate.endDate?.toISOString()).toBe(expectedEndDate);
+    expect(res?.compDate?.startDate?.toISOString()).toBe(expectedStartDate);
+    expect(res?.compDate?.endDate?.toISOString()).toBe(expectedEndDate);
   });
 
   it("finds the correct comp date for comp dates spanning over two different months", async () => {
@@ -19,7 +21,7 @@ describe("Airtribune", () => {
     const url = "https://airtribune.com/tennessee-paragliding-open-2023/";
     const res = await getAirtribuneComp(url);
 
-    expect(res?.compDate.startDate?.toISOString()).toBe(expectedStartDate);
-    expect(res?.compDate.endDate?.toISOString()).toBe(expectedEndDate);
+    expect(res?.compDate?.startDate?.toISOString()).toBe(expectedStartDate);
+    expect(res?.compDate?.endDate?.toISOString()).toBe(expectedEndDate);
   });
 });