+++ __tests__/get-wprs-civl.test.ts
@@ -1,3 +1,4 @@
+import { describe, it, expect } from "bun:test";
 import { getForecast } from "@/utils/get-forecast";
 import { getCivlcompsComp } from "@/utils/get-civl-comp";
 
@@ -8,7 +9,7 @@ describe("Get data for CIVL comps", () => {
     const url = "https://civlcomps.org/event/german-open-2023/participants";
     const res = await getCivlcompsComp(url);
 
-    expect(res.pilots.length).toBe(expectedNumOfPilots);
+    expect(res?.pilots.length).toBe(expectedNumOfPilots);
   }, 30000);
 
   it("should reject a comp that lies in the past", async () => {