+++ src/pages/index.tsx
@@ -13,6 +13,7 @@ import { ForecastView } from "@/components/ForecastView";
 import { useRouter } from "next/router";
 import { Footer } from "@/components/Footer";
 import { Element, scroller, animateScroll } from "react-scroll";
+import Link from "next/link";
 
 export type RecentQueriesProps = InferGetServerSidePropsType<
   typeof getServerSideProps
@@ -227,8 +228,9 @@ export default function Home(props: RecentQueriesProps) {
                       all.
                     </li>
                     <li>
-                      The calculation will become more accurate as the
-                      competition date approaches.
+                      The calculation will become{" "}
+                      <span className="text-green-500">more accurate</span> as
+                      the competition date approaches.
                     </li>
                     <li>
                       PWC events may currently not give correct results or not
@@ -239,6 +241,16 @@ export default function Home(props: RecentQueriesProps) {
                       It will not work for past events. It&apos;s a forecast!
                     </li>
                   </ul>
+                  <p className="mt-2 text-gray-500 dark:text-inherit">
+                    See the{" "}
+                    <Link
+                      className="text-green-500 hover:underline hover:decoration-green-500"
+                      href="/history"
+                    >
+                      history page
+                    </Link>{" "}
+                    for forecast accuracy statistics.
+                  </p>
                 </div>
               </Box>
             )}
@@ -283,20 +295,41 @@ export const getServerSideProps = async () => {
         id: true,
         compTitle: true,
         createdAt: true,
+        startDate: true,
+        endDate: true,
       },
       take: 50,
     });
 
-    const comps = data.map(({ createdAt, ...rest }) => {
+    const comps = data.map(({ createdAt, startDate, endDate, ...rest }) => {
       const now = new Date();
       const timeDiff = now.getTime() - createdAt.getTime(); // in milliseconds
 
       // Calculate time differences in hours and days
       const hoursDiff = Math.floor(timeDiff / (1000 * 60 * 60));
 
+      // Time till comp start
+      let daysTillCompStart: number | null = null;
+      let daysSinceCompEnd: number | null = null;
+      if (startDate != null) {
+        const diffInMilliseconds = startDate.getTime() - now.getTime();
+        daysTillCompStart = Math.ceil(
+          diffInMilliseconds / (1000 * 60 * 60 * 24),
+        );
+      }
+      // Time since comp end
+      if (endDate != null) {
+        const diffInMilliseconds = endDate.getTime() - now.getTime();
+        daysSinceCompEnd = Math.ceil(
+          diffInMilliseconds / (1000 * 60 * 60 * 24),
+        );
+      }
+
       return {
         ...rest,
         ageInHours: hoursDiff,
+        daysTillCompStart,
+        daysSinceCompEnd,
       };
     });