+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+### 2024-02-16
+
+**Improvements:**
+
+- History page with statistics added
+- PWCs should work again
+- Days till comp start added to table
+
 ### 2023-12-02
 
 **Improvements:**